+++ uni_modules/uv-overlay/components/uv-overlay/uv-overlay.vue
@@ -1,6 +1,7 @@
 <template>
 	<uv-transition
 	  :show="show"
+		mode="fade"
 	  custom-class="uv-overlay"
 	  :duration="duration"
 	  :custom-style="overlayStyle"
@@ -65,19 +66,20 @@
 		}
 	}
 </script>
-
 <style lang="scss" scoped>
-     $uv-overlay-top:0 !default;
-     $uv-overlay-left:0 !default;
-     $uv-overlay-width:100% !default;
-     $uv-overlay-height:100% !default;
-     $uv-overlay-background-color:rgba(0, 0, 0, .7) !default;
-	.uv-overlay {
-		position: fixed;
-		top:$uv-overlay-top;
-		left:$uv-overlay-left;
-		width: $uv-overlay-width;
-		height:$uv-overlay-height;
-		background-color:$uv-overlay-background-color;
-	}
+/* #ifndef APP-NVUE */
+$uv-overlay-top:0 !default;
+$uv-overlay-left:0 !default;
+$uv-overlay-width:100% !default;
+$uv-overlay-height:100% !default;
+$uv-overlay-background-color:rgba(0, 0, 0, .7) !default;
+.uv-overlay {
+	position: fixed;
+	top:$uv-overlay-top;
+	left:$uv-overlay-left;
+	width: $uv-overlay-width;
+	height:$uv-overlay-height;
+	background-color:$uv-overlay-background-color;
+}
+/* #endif */
 </style>