+++ uni_modules/uv-pick-color/components/uv-pick-color/props.js
@@ -1,10 +1,5 @@
 export default {
 	props: {
-		// 是否显示
-		show: {
-			type: Boolean,
-			default: false
-		},
 		// 颜色选择器初始颜色
 		color: {
 			type: Object,
@@ -20,7 +15,7 @@ export default {
 		// 是否允许点击遮罩关闭
 		closeOnClickOverlay: {
 			type: Boolean,
-			default: false
+			default: true
 		},
 		// 顶部标题
 		title: {