+++ uni_modules/uv-action-sheet/changelog.md
@@ -1,3 +1,5 @@
+## 1.0.2（2023-07-02）
+uv-action-sheet  由于弹出层uv-popup的修改，打开和关闭方法更改，详情参考文档：https://www.uvui.cn/components/actionSheet.html
 ## 1.0.1（2023-05-16）
 1. 优化组件依赖，修改后无需全局引入，组件导入即可使用
 2. 优化部分功能